Stats display system works again. Hud overlay supports text alignment, and avoids unnecessary font loading from disk.

+++ src/Skybolt/AircraftHud/HeadingRibbonModel.h
@@ -56,7 +56,7 @@ protected:
 	float clampXToHeadingRange(float x);
 
 	bool isMajorIncrement(float val);
-	std::string getTickLabel(float heading);
+	std::string getTickLabel(int heading);
 
 private:
 	HudDrawer* mDrawer;
